Compilable version with interfaces using crc32 to identify types, services, and calls
The rather ugly interface types are an attempt at adding type safety to C.

The organisation of the macros is a work in progress, especially the veneers, which
are unsafe, without any locking. Returning exceptions is also not supported.

Since the first services are values, rather than pointers to objects in the providers,
the macros are unreasonably inclined to that, unusual, use case.