Not quite broken.
The svc 0 at the end of the interrupt handler code seems to be essential
for the frame buffer to start up. Heaven knows why.

I'll try disabling all interrupt sources, to see if the problem still occurs.

Another thing to do is enable the mailbox interface interrupt, and use it,
instead of loops with yields in.