Returning from a thread function causes problems. Don't do it!
The el3 debug code has been enhanced to use the display, if it's been
initialised. It has the address of the frame buffer hard-coded, which
is bad.

The return address from a thread routine is not initialised; returns
just go to an unknown location.

The register initialisation code in initialise_new_thread can be
removed.