Altered the way a driver should return from an inter-map call.
Instead of returning from the __call_handler routine, which has some drawbacks*, driver
code can now call a routine associated with the type of the called object, returning
the parameters associated with the routine.

* Return types for different calls differ, the only easy type is a single value
* Difficult to preserve callee-saved registers, especially when throwing exceptions

At check in, returning multiple values or exceptions has not been tested (and the client
code doesn't deal with multiple value exception returns at all).

Probably the best place to start implementing and testing those features are in the
mailbox interface.