Progressing well, but about to break.
The VM implementation is progressing well, as I add support for assorted
system registers leading up to RISC OS running with the MMU enabled.

The implementation should be changed to use a VM interface provided by the
system driver, using the system call interface instead of separate SVCs, at
least for the relatively rare accesses to system registers.

The next change broke the execution, before anything useful happened. I
need to find out why, but keep this as a baseline.