Buildable with aarch64-linux-*-gcc
Modified the build script to deal with gcc that defaults to -fPIC (generating
a global offset table in the elf file).

Reorganised how memset is brought into the compiliation.

Removed an assumption about how the address of _start would be calculated.
This might be better done differently, but the EL3 and EL1 kernels run at
different addresses. PC-relative is still the way to go.

Still doesn't run without a modified qemu that loads the "bios" at 0 and
allocates fb memory on a 2MiB boundary.

Changed the build when generating for use with qemu; afaics it still doesn't
allow for an image file to be treated as an SD card at the hardware level.