Put 32-bit code for QEMU back into drivers/riscos.c
It was a mistake to move the 32-bit code for testing under qemu into the sd
driver: there's no emulated EMMC, so it just waited forever.

The 32-bit code gets started now, but fails very quickly. I'm checking this
in anyway, to test that the github code compiles (using
aarch64-linux-gnu-gcc-10), and the patched qemu-6.1.0.