Virtual interrupts are trivial
Next, trigger them from a real interrupt.

Timer interrupts, every half second, have the "guest OS" code
count up and occasionally switch to the partner thread to spend
some time "processing" (another count), so that the real IRQ
occurs either in Secure EL0 or NS EL1 (interrupts are disabled
in Secure EL1, EL2 and EL3, minimal time is spent in any of those
modes).

There will have to be an interrupt handler in riscos.c, or simply
a second thread that waits for the half second before raising the
virtual IRQ. Since the interrupt thread pre-empts whatever is
running, including the virtual machine, either should work.

The main consideration is that the thread that raises the
interrupt will have to be on (or signal) the core running the
virtual machine.